requirements.txt
cairocffi==1.7.1
CairoSVG==2.8.2
cffi==1.17.1
chess==1.11.2
colorama==0.4.6
contourpy==1.3.2
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
packaging==25.0
pillow==11.3.0
pycparser==2.22
pyparsing==3.2.3
python-chess==1.999
python-dateutil==2.9.0.post0
PyYAML==6.0.2
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tinycss2==1.4.0
tqdm==4.67.1
typing_extensions==4.14.1
webencodings==0.5.1
